Delete unnecessary files and folders
