+++ src/App.js
@@ -1,0 +1,13 @@
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <h1>My Covid App</h1>
+      </header>
+    </div>
+  );
+}
+
+export default App;
